+++ src/RampGenerator.h
@@ -66,6 +66,7 @@ class RampGenerator {
   int8_t move(int32_t move, const struct queue_end_s *queue);
   int8_t moveTo(int32_t position, const struct queue_end_s *queue);
   int8_t startRun(bool countUp);
+  inline void forceStop() { _ro.immediateStop(); }
   inline void initiateStop() { _ro.initiateStop(); }
   inline bool isStopping() {
     return _ro.isStopInitiated() && isRampGeneratorActive();