+++ tests/esp32_hw_based/seq_01.sh
@@ -2,7 +2,7 @@
 TTY=${1:-ttyUSB0}
 DEV="-d /dev/${TTY} -b 115200"
 
-CMD="M1 p7,0,0 H25000 A10000 f w1000 X pc R100 w100 R1000 W "
+CMD="M1 p7,0,0 H25000 A10000 f w1000 X pc R100 w100 W R1000 "
 PASS=">> M1: @1100 \\[1100\\]"
 
 LOG="$0.log"
@@ -20,6 +20,7 @@ echo
 #	echo FAIL $0 pulse counter mismatch
 #	exit 1
 #fi
+
 if [ `grep -c "$PASS" $LOG` -eq 1 ]
 then
 	echo PASS